core/chains: wip metis config; rename Optimism2 -> L2Suggested gas estimator (#6548)
* core/chains: wip metis config; rename Optimism2 -> L2Suggested gas estimator

* core/chains/evm/gas: remove OVM 1.0 gas estimator

* core/services/ocr: linter

* core/chains/evm/gas: cleanup and doc
