+++ core/config/chaintype.go
@@ -6,6 +6,7 @@ type ChainType string
 //nolint
 const (
 	ChainArbitrum ChainType = "arbitrum"
+	ChainMetis    ChainType = "metis"
 	ChainOptimism ChainType = "optimism"
 	ChainXDai     ChainType = "xdai"
 )
@@ -13,7 +14,7 @@ const (
 // IsValid returns true if the ChainType value is known or empty.
 func (c ChainType) IsValid() bool {
 	switch c {
-	case "", ChainArbitrum, ChainOptimism, ChainXDai:
+	case "", ChainArbitrum, ChainMetis, ChainOptimism, ChainXDai:
 		return true
 	}
 	return false
@@ -23,7 +24,7 @@ func (c ChainType) IsValid() bool {
 // used for log searching are different from calling block.number
 func (c ChainType) IsL2() bool {
 	switch c {
-	case ChainArbitrum, ChainOptimism:
+	case ChainArbitrum, ChainMetis, ChainOptimism:
 		return true
 
 	case ChainXDai:
