+++ core/web/resolver/evm_chain_config.go
@@ -17,6 +17,7 @@ type ChainType string
 
 const (
 	ChainTypeArbitrum ChainType = "ARBITRUM"
+	ChainTypeMetis    ChainType = "METIS"
 	ChainTypeOptimism ChainType = "OPTIMISM"
 	ChainTypeXDAI     ChainType = "XDAI"
 )
@@ -25,6 +26,8 @@ func ToChainType(s string) (ChainType, error) {
 	switch s {
 	case "arbitrum":
 		return ChainTypeArbitrum, nil
+	case "metis":
+		return ChainTypeMetis, nil
 	case "optimism":
 		return ChainTypeOptimism, nil
 	case "xdai":
@@ -38,6 +41,8 @@ func FromChainType(ct ChainType) string {
 	switch ct {
 	case ChainTypeArbitrum:
 		return "arbitrum"
+	case ChainTypeMetis:
+		return "metis"
 	case ChainTypeOptimism:
 		return "optimism"
 	case ChainTypeXDAI:
@@ -52,8 +57,8 @@ type GasEstimatorMode string
 const (
 	GasEstimatorModeBlockHistory GasEstimatorMode = "BLOCK_HISTORY"
 	GasEstimatorModeFixedPrice   GasEstimatorMode = "FIXED_PRICE"
-	GasEstimatorModeOptimism     GasEstimatorMode = "OPTIMISM"
 	GasEstimatorModeOptimism2    GasEstimatorMode = "OPTIMISM2"
+	GasEstimatorModeL2Suggested  GasEstimatorMode = "L2_SUGGESTED"
 )
 
 func ToGasEstimatorMode(s string) (GasEstimatorMode, error) {
@@ -62,10 +67,10 @@ func ToGasEstimatorMode(s string) (GasEstimatorMode, error) {
 		return GasEstimatorModeBlockHistory, nil
 	case "FixedPrice":
 		return GasEstimatorModeFixedPrice, nil
-	case "Optimism":
-		return GasEstimatorModeOptimism, nil
 	case "Optimism2":
 		return GasEstimatorModeOptimism2, nil
+	case "L2Suggested":
+		return GasEstimatorModeL2Suggested, nil
 	default:
 		return "", errors.New("invalid gas estimator mode")
 	}
@@ -77,10 +82,10 @@ func FromGasEstimatorMode(gsm GasEstimatorMode) string {
 		return "BlockHistory"
 	case GasEstimatorModeFixedPrice:
 		return "FixedPrice"
-	case GasEstimatorModeOptimism:
-		return "Optimism"
 	case GasEstimatorModeOptimism2:
 		return "Optimism2"
+	case GasEstimatorModeL2Suggested:
+		return "L2Suggested"
 	default:
 		return strings.ToLower(string(gsm))
 	}
